Add files via upload
A Simple Linear Programming (LP) Model to optimize the number of each type of phone produced such that over all cost is minimized. Toward this  docplex library is used which is library used to solve optimization problems using CPLEX engine. CPLEX is a solver developed by IBM for mathematical programming.